// frontend/my-yen2-app/src/api/models/NavigationItemUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type NavigationItemUpdate = {
    title?: (string | null);
    url?: (string | null);
    icon_url?: (string | null);
    description?: (string | null);
    order_index?: (number | null);
};